requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
mongoengine==0.29.1
pymongo==4.12.0
sqlparse==0.5.3
